bug fix for pen demo.

+++ src/java/rexos/mas/service_agent/behaviours/GetPartsInfoResponse.java
@@ -180,7 +180,7 @@ public class GetPartsInfoResponse extends ReceiveOnceBehaviour {
 								ServiceStep.sort(serviceSteps));
 
 				ScheduleData scheduleData;
-				int nextStartTime = productStep.getScheduleData().getStartTime();
+				long nextStartTime = productStep.getScheduleData().getStartTime();
 				for(ServiceStep serviceStep : parameterizedSteps) {
 					scheduleData = serviceStep.getScheduleData();
 					scheduleData.setStartTime(nextStartTime);
